TASK: refactor reference(s) components to be DRYer
